@@ -1,15 +1,14 @@
 package com.fw.cobranca.api;
 
-import com.fw.cobranca.domain.*;
-import com.fw.cobranca.domain.dto.EstabelecimentoDTO;
-import com.fw.cobranca.service.*;
+import com.fw.cobranca.domain.Caixa;
+import com.fw.cobranca.service.CaixaService;
 import com.fw.cobranca.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.*;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/v1/cobranca")
