@@ -1,11 +1,14 @@
 package com.fw.cobranca.api;
 
-import com.fw.cobranca.domain.*;
-import com.fw.cobranca.service.*;
+import com.fw.cobranca.domain.Usuario;
+import com.fw.cobranca.service.LoginService;
 import com.fw.cobranca.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
