@@ -2,7 +2,7 @@ package com.fw.cobranca.api;
 
 import com.fw.cobranca.domain.Parcela;
 import com.fw.cobranca.domain.dto.ParcelaHomeDTO;
-import com.fw.cobranca.service.*;
+import com.fw.cobranca.service.ParcelaService;
 import com.fw.cobranca.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -53,4 +53,16 @@ public class ParcelaController implements Util {
             return new ResponseEntity<>(cobrancaController.mapErro("PUT", ex.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
+
+    @GetMapping("/parcela/verifica")
+    public ResponseEntity getParcelaPorEmprestimo(@RequestParam("id_emprestimo") Integer idEmprestimo,
+                                                  @RequestParam("numero_parcela") Integer numeroParcela) {
+        try {
+            return ResponseEntity.ok(parcelaService.verificaParcelaAnteriorEmAberto(idEmprestimo, numeroParcela));
+        } catch (Exception ex) {
+            return new ResponseEntity<>(cobrancaController.mapErro("GET", ex.getMessage()), HttpStatus.BAD_REQUEST);
+        }
+    }
+
+
 }
