@@ -1,14 +1,17 @@
 package com.fw.cobranca.api;
 
-import com.fw.cobranca.domain.*;
-import com.fw.cobranca.service.*;
+import com.fw.cobranca.domain.Usuario;
+import com.fw.cobranca.service.UsuarioService;
 import com.fw.cobranca.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/cobranca")
