@@ -2,9 +2,9 @@ package com.fw.cobranca.domain;
 
 import com.fw.cobranca.util.Util;
 import lombok.Data;
+
 import javax.persistence.*;
 import java.text.ParseException;
-import java.util.List;
 import java.util.Map;
 
 @Entity
@@ -15,6 +15,7 @@ public class Estabelecimento implements Util {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private Integer idUsuario;
+    private Integer idRegiao;
     private String cnpj;
     private String nome;
     private String endereco;
@@ -34,6 +35,7 @@ public class Estabelecimento implements Util {
     public Estabelecimento(Map map) throws ParseException {
         this.id = toInt(map.get("id"));
         this.idUsuario = toInt(map.get("idUsuario"));
+        this.idRegiao = toInt(map.get("idRegiao"));
         this.cnpj = toStr(map.get("cnpj"));
         this.nome = toStr(map.get("nome"));
         this.endereco = toStr(map.get("endereco"));
