@@ -1,8 +1,6 @@
 package com.fw.cobranca.domain.dto;
 
 import com.fw.cobranca.domain.Emprestimo;
-import com.fw.cobranca.domain.Estabelecimento;
-import com.fw.cobranca.domain.TipoEmprestimo;
 import com.fw.cobranca.util.Util;
 
 import java.util.List;
@@ -19,5 +17,6 @@ public class HomeDTOAdministrador implements Util {
     public List<Emprestimo> emprestimosEmAtraso;
     public List<Emprestimo> emprestimosAVencer;
     public Iterable<ParcelaHomeDTO> parcelasAguardandoAnalise;
+    public Integer quantidadeNotificacaoNaoLida;
 
 }
