@@ -1,8 +1,6 @@
 package com.fw.cobranca.domain.dto;
 
 import com.fw.cobranca.domain.Emprestimo;
-import com.fw.cobranca.domain.Estabelecimento;
-import com.fw.cobranca.domain.TipoEmprestimo;
 import com.fw.cobranca.util.Util;
 
 import java.util.List;
@@ -13,6 +11,13 @@ public class HomeDTOColaborador implements Util {
     public Double saldo;
 
     public Iterable<Emprestimo> emprestimosAguardandoAnalise;
+    public Iterable<Emprestimo> emprestimosEmAnalise;
     public List<Emprestimo> emprestimosEmAtraso;
     public List<Emprestimo> emprestimosAVencer;
+    public Integer quantidadeNotificacaoNaoLida;
+
+    public Double totalEmprestimosAguardandoAnalise;
+    public Double totalEmprestimosEmAnalise;
+    public Double totalParcelasEmAtraso;
+    public Double totalParcelasAVencer;
 }
